+++ camerakit/utils/enumerator.py
@@ -1,6 +1,10 @@
+import os
+os.environ["OPENCV_LOG_LEVEL"] = "SILENT"  # Suppress OpenCV warnings.
+
 import platform
 from dataclasses import dataclass
 from typing import List, Optional, Tuple
+
 
 import cv2
 
@@ -9,6 +13,7 @@ from .common import (
     get_camera_hardware_macos,
     get_camera_properties,
     get_logger,
+    get_preferred_capture_backend,
     suppress_stderr,
 )
 
@@ -145,6 +150,7 @@ class CameraEnumerator:
         self.aspect_ratios = aspect_ratios
         self.common_widths = common_widths
         self.codecs = codecs
+        self.backend = get_preferred_capture_backend()
 
     def list(self) -> List[CameraMetadata]:
         """List all available cameras with metadata and supported settings.
@@ -158,7 +164,7 @@ class CameraEnumerator:
 
             if system == "Darwin":  # macOS
                 for cam in get_camera_hardware_macos():
-                    cap = cv2.VideoCapture(cam["id"])
+                    cap = cv2.VideoCapture(cam["id"], self.backend)
                     if not cap.isOpened():
                         get_logger().error(
                             f"Failed to open camera {cam['id']} on macOS."
@@ -183,6 +189,7 @@ class CameraEnumerator:
                         aspect_ratios=self.aspect_ratios,
                         common_widths=self.common_widths,
                         codecs=self.codecs,
+                        backend=self.backend,
                     )
                     if camera:
                         discovered.append(camera)